# public/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "RESTfull api for Tiler",
    "version": "1.0.0"
  },
  "host": "api.tiler.com",
  "basePath": "\/",
  "schemes": [
    "http"
  ],
  "paths": {
    "\/oauth\/token": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Authorization",
        "produces": [
          "application\/json"
        ],
        "parameters": [
          {
            "name": "grant_type",
            "in": "query",
            "description": "Oauth2.0 grand type - password",
            "required": true,
            "type": "string"
          },
          {
            "name": "client_id",
            "in": "query",
            "description": "Oauth client id",
            "required": true,
            "type": "string"
          },
          {
            "name": "client_secret",
            "in": "query",
            "description": "Oauth client secret",
            "required": true,
            "type": "string"
          },
          {
            "name": "username",
            "in": "query",
            "description": "User login",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "query",
            "description": "User password",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success authorization"
          }
        }
      }
    },
    "\/categories": {
      "get": {
        "tags": [
          "Categories"
        ],
        "summary": "Get all categories",
        "produces": [
          "application\/json"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "\/users\/{user_id}\/collections": {
      "get": {
        "tags": [
          "Collections"
        ],
        "summary": "Get user's collections",
        "produces": [
          "application\/json"
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "User id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success get post detail"
          }
        }
      },
      "post": {
        "tags": [
          "Collections"
        ],
        "summary": "Create new collection",
        "consumes": [
          "multipart\/form-data"
        ],
        "produces": [
          "application\/json"
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "User id",
            "required": true,
            "type": "string"
          },
          {
            "name": "name",
            "in": "query",
            "description": "Collection's name",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Success create new collection"
          },
          "403": {
            "description": "Access denied"
          },
          "422": {
            "description": "Validation errors"
          }
        },
        "security": [
          {
            "passport": []
          }
        ]
      }
    },
    "\/posts": {
      "get": {
        "tags": [
          "Posts"
        ],
        "summary": "Get all posts",
        "produces": [
          "application\/json"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "\/posts\/{post_id}": {
      "get": {
        "tags": [
          "Posts"
        ],
        "summary": "Post detail",
        "produces": [
          "application\/json"
        ],
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "description": "Post id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success get post detail"
          }
        }
      }
    },
    "\/users\/{user_id}\/posts": {
      "get": {
        "tags": [
          "Posts"
        ],
        "summary": "Get posts by user",
        "produces": [
          "application\/json"
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "User id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success get post detail"
          }
        }
      },
      "post": {
        "tags": [
          "Posts"
        ],
        "summary": "Publish new post",
        "consumes": [
          "multipart\/form-data"
        ],
        "produces": [
          "application\/json"
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "User id",
            "required": true,
            "type": "string"
          },
          {
            "name": "title",
            "in": "query",
            "description": "Title of post",
            "required": true,
            "type": "string"
          },
          {
            "name": "text",
            "in": "query",
            "description": "Post's text",
            "required": true,
            "type": "string"
          },
          {
            "name": "preview",
            "in": "formData",
            "description": "Post's preview",
            "required": false,
            "type": "string"
          },
          {
            "name": "category_id",
            "in": "query",
            "description": "Post's category id",
            "required": false,
            "type": "integer"
          },
          {
            "name": "collection_id",
            "in": "query",
            "description": "Post's collection id",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "201": {
            "description": "Success create new post"
          },
          "403": {
            "description": "Access denied"
          },
          "422": {
            "description": "Validation errors"
          }
        },
        "security": [
          {
            "passport": []
          }
        ]
      }
    },
    "\/users": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Registration new user",
        "produces": [
          "application\/json"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "User first name",
            "required": true,
            "type": "string"
          },
          {
            "name": "surname",
            "in": "query",
            "description": "User last name",
            "required": false,
            "type": "string"
          },
          {
            "name": "email",
            "in": "query",
            "description": "User email",
            "required": true,
            "type": "string"
          },
          {
            "name": "avatar",
            "in": "query",
            "description": "User avatar in base64 encode",
            "required": false,
            "type": "string"
          },
          {
            "name": "password",
            "in": "query",
            "description": "User password",
            "required": true,
            "type": "string"
          },
          {
            "name": "password_confirmation",
            "in": "query",
            "description": "User confirmation password",
            "required": true,
            "type": "string"
          },
          {
            "name": "terms",
            "in": "query",
            "description": "check terms of use",
            "required": true,
            "type": "boolean"
          }
        ],
        "responses": {
          "201": {
            "description": "Success create new user"
          },
          "422": {
            "description": "Validation errors"
          }
        }
      }
    },
    "\/users\/{id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "User detail",
        "produces": [
          "application\/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success get user detail"
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Update user",
        "produces": [
          "application\/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User id",
            "required": true,
            "type": "string"
          },
          {
            "name": "name",
            "in": "query",
            "description": "User first name",
            "required": false,
            "type": "string"
          },
          {
            "name": "surname",
            "in": "query",
            "description": "User last name",
            "required": false,
            "type": "string"
          },
          {
            "name": "email",
            "in": "query",
            "description": "User email",
            "required": false,
            "type": "string"
          },
          {
            "name": "avatar",
            "in": "query",
            "description": "User avatar in base64 encode",
            "required": false,
            "type": "string"
          },
          {
            "name": "password",
            "in": "query",
            "description": "New user password",
            "required": false,
            "type": "string"
          },
          {
            "name": "password_confirmation",
            "in": "query",
            "description": "Confirmation new user password",
            "required": false,
            "type": "string"
          },
          {
            "name": "current_password",
            "in": "query",
            "description": "Current user password",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success user update"
          },
          "400": {
            "description": "authorization token is required"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Access denied"
          },
          "422": {
            "description": "Validation errors"
          }
        },
        "security": [
          {
            "passport": []
          }
        ]
      }
    }
  },
  "definitions": {},
  "securityDefinitions": {
    "passport": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}